requirements.txt
Flask>=3.0.0
requests>=2.32.0
trafilatura>=1.7.0
pypdf>=5.0.0
python-dotenv>=1.0.1
openai>=1.40.0
google-generativeai>=0.7.2
